+++ src/models/Address.php
@@ -539,7 +539,7 @@ class Address extends Model
         $this->trigger(self::EVENT_DEFINE_ADDRESS_LINES, $event);
 
         if ($sanitize) {
-            array_walk($event->addressLines, function(&$value, &$key) {
+            array_walk($event->addressLines, function(&$value) {
                 $value = Craft::$app->getFormatter()->asText($value);
             });
         }