+++ tests/fixtures/AddressesFixture.php
@@ -8,7 +8,7 @@
 namespace craftcommercetests\fixtures;
 
 use craft\commerce\records\Address;
-use craft\test\Fixture;
+use craft\test\ActiveFixture;
 
 /**
  * Addresses Fixture
@@ -16,7 +16,7 @@ use craft\test\Fixture;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.1.4
  */
-class AddressesFixture extends Fixture
+class AddressesFixture extends ActiveFixture
 {
     /**
      * @inheritdoc