+++ tests/unit/services/PaymentCurrenciesTest.php
@@ -20,7 +20,7 @@ use UnitTester;
  * Payment Currencies Test
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
- * @since 3.x
+ * @since 3.2.14
  */
 class PaymentCurrenciesTest extends Unit
 {